Refactor: rename files and create summary folder for confirm
