+++ src/app/address/controllers/address/manual.js
@@ -2,12 +2,13 @@ const BaseController = require("hmpo-form-wizard").Controller;
 
 const {
   validateHouseNumberAndName,
-} = require("../validators/addressValidator");
+} = require("../../validators/addressValidator");
 
 class AddressController extends BaseController {
   getValues(req, res, callback) {
     super.getValues(req, res, (err, values) => {
       const address = req.sessionModel.get("address");
+
       values.addressPostcode =
         address?.postalCode || req.sessionModel.get("addressPostcode");
 
@@ -58,7 +59,15 @@ class AddressController extends BaseController {
         address.postalCode = req.sessionModel.get("addressPostcode");
       }
 
-      req.sessionModel.set("address", address);
+      const currentWizard = req?.sessionModel?.options?.key;
+      req.sessionModel.set("address", address); // set for /edit routes
+
+      // set for /confirm
+      if (currentWizard.endsWith("previous")) {
+        req.sessionModel.set("previousAddress", address);
+      } else {
+        req.sessionModel.set("currentAddress", address);
+      }
       callback();
     });
   }
