+++ src/views/summary/confirm.html
@@ -4,10 +4,7 @@
 {% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
 {% from "hmpo-radios/macro.njk" import hmpoRadios %}
 
-
-
 {% set hmpoPageKey = "address-confirm" %}
-
 
 {% block mainContent %}
